rpcap: clarify what the linktype preference does.
"Received" could mean "received by the rpcap client" or "received as
part of the capturing process by Wireshark".  The former would mean "no
capture is happening", so there would be no packets *to* interpret using
that preference; use "captured" to clarify that it's the latter (which
could mean it was dropped as part of the capture process, or it could
just mean we started capturing in the middle of the remote capture
process).