Fix issues with tcptrace graph
1. In order to keep error bars and scatterplot in-sync for segment and
SACK graphs, tell QCustomPlot that the data is already sorted when
adding it to scatter plots. [ Otherwise it will re-sort the data and
potentially get out-of-sync with the error bars graphs ]

2. Clear and hide error bars plottables in fillGraph() similar to what
is done for other graphs.  [ otherwise QCustomPlot may attempt to draw
error bars on graphs where they are not relevant ]

3. Do not mark window update packets as duplicate acks.