ieee80211: Split out the SS Allocation field to two fields.
The SS Allocation field is listed in the spec as two separate fields. The
Starting Spatial Stream number and the Number of Spatial Streams. While
the spec treasts this as one 6-bit field it is more useful to display it as
two 3-bit fields so users can search on each separately.

Change-Id: I7a510a66f4500ac9fe065845ade9f5c3aaaf2184