packet-wow: Correct protocol_version field
Fixes a very small part of #11398.

The fields were incorrectly identified as error fields because Mangos,
the most popular open source emulator in 2009 incorrectly identified it
as such.

Reverse engineering of the client revealed that the fields were protocol
versions.

As well as misidentifying the field, the WOW_SERVER_TO_CLIENT version
also read the wrong field. The actual error field is the one that comes
after the protocol version field.

Correct packets are in Ember
WOW_CLIENT_TO_SERVER
https://github.com/EmberEmu/Ember/blob/03c130d3d6276e7032fc9e13c9d287ea7c6ed536/src/login/grunt/client/LoginChallenge.h#L39

WOW_SERVER_TO_CLIENT
https://github.com/EmberEmu/Ember/blob/03c130d3d6276e7032fc9e13c9d287ea7c6ed536/src/login/grunt/server/LoginChallenge.h#L33

As well as the Wiki
https://wowdev.wiki/Packets/Login/Vanilla#Challenge_packets