Add a code of conduct.
After reviewing several codes of conduct, the KDE CoC[1][2] seems to
match the goals and personality of our community the best. Adapt it to
suit our needs. Name the file CODE_OF_CONDUCT.adoc, which seems to be
the most common base name. Note that it must be kept in sync with
https://www.wireshark.org/code-of-conduct.html.

[1]https://kde.org/code-of-conduct/
[2]https://invent.kde.org/websites/kde-org/-/blob/master/content/code-of-conduct.md