RPCoRDMA: fix read chunk reassembly if reply chunk is present
Use correct segment type given by the reassembly message id when
dealing with multiple chunks in the same frame.

When an NFS WRITE is sent with a read chunk having multiple
segments and it also includes a reply chunk, using the wrong
segment type could lead to reassemble the message prematurely
on the last read response for the first read chunk segment.
The message should be reassembled in the last read response
for the last segment.