dof: clean up adding the session key to the dissection.
Let the libwireshark code format the value of the session key, by
setting its display format to SEP_COLON.

Use proto_tree_add_bytes_with_length(), so that the session key becomes
the value of the field; otherwise, the length of the key isn't
specified.

This means we don't need to call bytestring_to_str() (the code in
epan/proto.c calls it for us).

This also plugs a leak of the string generated by bytestring_to_str().