Plug a memory leak.
If the user aborted the process of exporting packets, if we're writing
to a temporary file, we unlink the file, but we don't free the
g_mallocated name of the file, so it's leaked.  Free it.

(cherry picked from commit 54508703b0ac4d85df4104f6114a247d68c42b8e)