Remove a duplicate unlink.
In cf_export_specified_packets(), if the loop processing the packets
fails, we're going to go to the failure code, which will unlink the file
to which we were writing if we were writing to a temporary file, so we
don't need to unlink it before going there.

While we're at it, note why we don't report any error from
wtap_dump_close() in that case.