Plug another leak.
If cf_export_specified_packets() succeeds, and it wrote to a temporary
file, it leaks the name of the file to which it was writing.  Free that
after we've renamed that file on top of the target file (safe save).


(cherry picked from commit 8ca86b29bfa4780b17e6d15ab1f382f6f79a5065)