DVB-S2-BB: Add preference for default mode adaptation format
DVB Baseband Frames have four possible mode adaptation formats, and
the check to distinguish them depends on only a CRC-8, so it's possible
to have ambiguity, particularly in a large capture. Add a preference to
for preferred format, defaulting to the SatLabs L.3 mode, since that's
the most common format I've seen in captures or in datasheets for
BBFrames over UDP or RTP. Also make L.3 the second highest preference
when something else is set to the default.