ieee80211_radio: Use calculated bitrate if not supplied
For 802.11n if the bitrate is not supplied then the calculated bitrate is used. This change does the same for 11ac.

Sniffer traces taken on recent versions of Macos no longer supply the bitrate for 11ac frames in the RADIOTAP header, this change allows the wireless timeline to work with these traces.

Fixes #17419.

(backported from commit 5202119239f76af3df20540b35c618d7095c99c6)