ipsec: refactor the decryption and dissection of ESP packets
Adding full support for AEAD ciphers like AES-GCM (including the
verification of the ICV) turned out to be difficult with the
current implementation of the ipsec dissector, because it does not
separate the dissection and decryption steps well enough and has
too many special cases depending on the crypto algorithms.

From a dissector's viewpoint there shouldn't be much difference
between an authenticated encryption method and a combination of
a classical encryption method with an authentication method.
What matters is how the data is structured, so much how it is
calculated.

       HEADER || IV || ENCRYPTED DATA || ICV      (Frame Data)
                               |
                               v
                       DECRYPTED DATA             (Decrypted Data)

This commit tries to refactor the implementation with the goal to
minimize the differences between the different crypto operations,
in particular their operation modes (like AES-CBC,AES-CTR,AES-GCM).
It follows the example of the isakmp dissector for IKEv2 packets,
which already has a functional AES-GCM support.

The most significant changes are:

- Display the IV and ICV as part of the original Frame Data, not
  as part of the Decrypted Data.

  - Display the location of the encrypted data, together with
    information about encryption and authentication algorithms.

  - Use gcry_cipher_setiv() to set the IV for AES-CBC instead of
    copying the IV into the decryption buffer as a prefix which
    subsequently gets discarded.

  - Don't copy the ICV into the decryption buffer where it gets
    "decrypted" accidentally and needs to be restored afterwards.

- Strip the nonce from the encryption key for AES-GCM and AES-CTR
  at an early stage, to reduce special treatment for those modes
  due to the different key lengths.

- Add some missing dissection tree items to get full coverage
  of all bytes in the Frame Data and Decrypted Data.

- Don't report dissector bugs to stderr. Instead, use the
  REPORT_DISSECTOR_BUG() macro which will raise an exception.
  (If the WIRESHARK_ABORT_ON_DISSECTOR_BUG environment variable
  is set, the program will call abort() instead, to make it easier
  to get a stack trace.)

With these changes, AES-GCM encrypted payloads now get dissected
correctly after decryption, provided the ICV length is specified
correctly. The ICV verification is still missing, it will be added
in a followup commit.